FEAT/DOCS: Adjusted the views to ensure that they are secured and only display the right elements applicable to the type of user using the site. This ensures non-registered users or non-superusers cannot bypass the site links.

+++ products/views.py
@@ -3,6 +3,8 @@ from django.contrib import messages
 # to allow for search criteria in product name OR description
 from django.db.models import Q, functions
 
+from django.contrib.auth.decorators import login_required
+
 from .models import Product, Category
 from .forms import ProductAdminForm
 
@@ -75,9 +77,14 @@ def product_info(request, product_id):
     return render(request, 'products/product_info.html', context)
 
 
+@login_required
 def add_product(request):
     """ A view to allow StudyHaven superuser and owner to add
     a product to the shop. """
+    if not request.user.is_superuser:
+        messages.error(request, 'Only the site owners can do that!')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         admin_form = ProductAdminForm(request.POST, request.FILES)
         if admin_form.is_valid():
@@ -99,9 +106,14 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ A view to allow StudyHaven superuser and owner to edit
     a product already in the shop. """
+    if not request.user.is_superuser:
+        messages.error(request, 'Only the site owners can do that!')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
 
     if request.method == 'POST':
@@ -127,8 +139,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ A view to delete the product from the StudyHaven shop. """
+    if not request.user.is_superuser:
+        messages.error(request, 'Only the site owners can do that!')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, f'{product.name} has now been deleted from the StudyHaven shop!')
